sys/newshell: moved new shell functionality to sys directory
Detached new shell functionality from app and moved it to a
package in sys. In the future it will replace original shell.
