+++ apps/newshell/src/main.c
@@ -22,38 +22,12 @@
 #include "sysinit/sysinit.h"
 #include "os/os.h"
 
-#include "shell_priv.h"
-#include "shell.h"
-
-#define SHELL_TASK_PRIO         (3)
-#define SHELL_STACK_SIZE        (512)
-static struct os_task shell_task;
-static os_stack_t shell_task_stack[SHELL_STACK_SIZE];
-#define SHELL_PROMPT "shell> "
-
-static void
-init_task(void)
-{
-    /* Initialize the task */
-    os_task_init(&shell_task, "shell", shell, NULL,
-                 SHELL_TASK_PRIO, OS_WAIT_FOREVER,
-                 shell_task_stack, SHELL_STACK_SIZE);
-    shell_init(SHELL_PROMPT);
-#if MYNEWT_VAL(SHELL_OS_MODULE)
-    shell_os_register(shell_register);
-#endif
-#if MYNEWT_VAL(SHELL_PROMPT_MODULE)
-    shell_prompt_register(shell_register);
-#endif
-}
-
 int
 main(int argc, char **argv)
 {
     int rc;
 
     sysinit();
-    init_task();
 
     while (1) {
         os_eventq_run(os_eventq_dflt_get());
