+++ sys/newshell/src/shell_priv.h
