+++ sys/newshell/syscfg.yml
@@ -1,0 +1,49 @@
+# Licensed to the Apache Software Foundation (ASF) under one
+# or more contributor license agreements.  See the NOTICE file
+# distributed with this work for additional information
+# regarding copyright ownership.  The ASF licenses this file
+# to you under the Apache License, Version 2.0 (the
+# "License"); you may not use this file except in compliance
+# with the License.  You may obtain a copy of the License at
+#
+#  http://www.apache.org/licenses/LICENSE-2.0
+#
+# Unless required by applicable law or agreed to in writing,
+# software distributed under the License is distributed on an
+# "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY
+# KIND, either express or implied.  See the License for the
+# specific language governing permissions and limitations
+# under the License.
+#
+
+# Package: sys/newshell 
+
+syscfg.defs:
+    SHELL_TASK:
+        description: 'Controls whether shell is enabled or not.'
+        value: 0
+
+    SHELL_CMD_HELP:
+        description: 'Include help information for shell commands'
+        value: 1
+
+    SHELL_OS_MODULE:
+        description: 'Include shell os module'
+        value: 1
+    SHELL_PROMPT_MODULE:
+        description: 'Include shell prompt module'
+        value: 1
+
+    CONSOLE_RTT:
+        description: 'Enable console over RTT'
+        value: 0
+    CONSOLE_UART:
+        description: 'Enable console over UART'
+        value: 0
+
+    CONSOLE_UART_BAUD:
+        description: 'Console UART baud rate.'
+        value: '115200'
+    CONSOLE_UART_FLOW_CONTROL:
+        description: 'Console UART flow control.'
+        value: 'UART_FLOW_CTL_NONE'
